commited by Abner

+++ ze-delivery-spring-boot-back-end-java/src/main/java/com/org/ze_delivery_spring_boot_back_end_java/models/dtos/responses/CoverageAreaResponse.java
@@ -1,5 +1,6 @@
 package com.org.ze_delivery_spring_boot_back_end_java.models.dtos.responses;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 import com.org.ze_delivery_spring_boot_back_end_java.models.enums.CoverageAreaType;
@@ -13,5 +14,5 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 public class CoverageAreaResponse {
 	private CoverageAreaType type; 
-	private List<List<List<Long>>> coordinates;
+	private List<List<List<BigDecimal>>> coordinates;
 }
